% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2d_dist.R
\name{print.rd2d.dist}
\alias{print.rd2d.dist}
\title{Print Method for 2D Local Polynomial RD Estimation (Distance-Based)}
\usage{
\method{print}{rd2d.dist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rd2d.dist}, returned by \code{\link{rd2d.dist}}.}

\item{...}{Additional arguments passed to the method (currently ignored).}
}
\value{
No return value. This function is called for its side effects: it prints the \code{\link{rd2d.dist}} results.
}
\description{
Prints the results of a 2D local polynomial regression discontinuity (RD) estimation using distance-based evaluation, as obtained from \code{\link{rd2d.dist}}.
}
\seealso{
\code{\link{rd2d.dist}} for estimation using distance-based methods in 2D local polynomial RD designs.

Supported methods: \code{\link{print.rd2d.dist}}, \code{\link{summary.rd2d.dist}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu} \cr
Rocío Titiunik, Princeton University. \email{titiunik@princeton.edu} \cr
Ruiqi Rae Yu, Princeton University. \email{rae.yu@princeton.edu}
}
