% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2d_distance.R
\name{print.rd2d.distance}
\alias{print.rd2d.distance}
\title{Print Method for 2D Local Polynomial RD Estimation (Distance-Based)}
\usage{
\method{print}{rd2d.distance}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rd2d.distance}, returned by \code{\link{rd2d.distance}}.}

\item{...}{Additional arguments passed to the method (currently ignored).}
}
\value{
No return value. This function is called for its side effects: it prints the \code{\link{rd2d.distance}} results.
}
\description{
Prints the results of a 2D local polynomial regression discontinuity (RD) estimation using distance-based evaluation, as obtained from \code{\link{rd2d.distance}}.
}
\seealso{
\code{\link{rd2d.distance}} for estimation using distance-based methods in 2D local polynomial RD designs.

Supported methods: \code{\link{print.rd2d.distance}}, \code{\link{summary.rd2d.distance}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com} \cr
Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu} \cr
Ruiqi Rae Yu, Princeton University. \email{raeyuuuu@gmail.com}
}
