% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2d.R
\name{print.rd2d}
\alias{print.rd2d}
\title{Print Method for 2D Local Polynomial RD Estimation}
\usage{
\method{print}{rd2d}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rd2d}, returned by \code{\link{rd2d}}.}

\item{...}{Additional arguments passed to the method (currently ignored).}
}
\value{
No return value. This function is called for its side effects, which are to print the \code{\link{rd2d}} results.
}
\description{
Prints the results of a 2D local polynomial regression discontinuity (RD) estimation, as obtained from \code{\link{rd2d}}.
}
\seealso{
\code{\link{rd2d}} for conducting 2D local polynomial RD estimation.

Supported methods: \code{\link{print.rd2d}}, \code{\link{summary.rd2d}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu} \cr
Rocío Titiunik, Princeton University. \email{titiunik@princeton.edu} \cr
Ruiqi Rae Yu, Princeton University. \email{rae.yu@princeton.edu}
}
