% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbw2d_dist.R
\name{print.rdbw2d.dist}
\alias{print.rdbw2d.dist}
\title{Print Method for Bandwidth Selection (Distance-Based) in 2D Local Polynomial RD Design}
\usage{
\method{print}{rdbw2d.dist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rdbw2d.dist}, returned by \code{\link{rdbw2d.dist}}.}

\item{...}{Additional arguments passed to the method (currently ignored).}
}
\value{
No return value. This function is called for its side effects: it prints summary information of \code{\link{rdbw2d.dist}}.
}
\description{
Print method for displaying summary information from distance-based bandwidth selection in 2D local polynomial regression discontinuity (RD) designs, as produced by \code{\link{rdbw2d.dist}}.
}
\seealso{
\code{\link{rdbw2d.dist}} for distance-based bandwidth selection in 2D local polynomial RD design.

Supported methods: \code{\link{print.rdbw2d.dist}}, \code{\link{summary.rdbw2d.dist}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu} \cr
Rocío Titiunik, Princeton University. \email{titiunik@princeton.edu} \cr
Ruiqi Rae Yu, Princeton University. \email{rae.yu@princeton.edu}
}
