% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbw2d_distance.R
\name{print.rdbw2d.distance}
\alias{print.rdbw2d.distance}
\title{Print Method for Bandwidth Selection (Distance-Based) in 2D Local Polynomial RD Design}
\usage{
\method{print}{rdbw2d.distance}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rdbw2d.distance}, returned by \code{\link{rdbw2d.distance}}.}

\item{...}{Additional arguments passed to the method (currently ignored).}
}
\value{
No return value. This function is called for its side effects: it prints summary information of \code{\link{rdbw2d.distance}}.
}
\description{
Print method for displaying summary information from distance-based bandwidth selection in 2D local polynomial regression discontinuity (RD) designs, as produced by \code{\link{rdbw2d.distance}}.
}
\seealso{
\code{\link{rdbw2d.distance}} for distance-based bandwidth selection in 2D local polynomial RD design.

Supported methods: \code{\link{print.rdbw2d.distance}}, \code{\link{summary.rdbw2d.distance}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com} \cr
Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu} \cr
Ruiqi Rae Yu, Princeton University. \email{raeyuuuu@gmail.com}
}
