% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbw2d.R
\name{print.rdbw2d}
\alias{print.rdbw2d}
\title{Print Method for Bandwidth Selection for 2D Local Polynomial RD Design}
\usage{
\method{print}{rdbw2d}(x, ...)
}
\arguments{
\item{x}{Class \code{rdbw2d} objects, obtained by calling \code{\link{rdbw2d}}.}

\item{...}{Additional arguments passed to the method (currently ignored).}
}
\value{
No return value, called to print \code{\link{rdbw2d}} results.
}
\description{
The print method for bandwidth selection for 2D local polynomial RD design
}
\seealso{
\code{\link{rdbw2d}} for bandwidth selection for 2D local polynomial RD design

Supported methods: \code{\link{print.rdbw2d}}, \code{\link{summary.rdbw2d}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu} \cr
Rocío Titiunik, Princeton University. \email{titiunik@princeton.edu} \cr
Ruiqi Rae Yu, Princeton University. \email{rae.yu@princeton.edu}
}
