% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2d_methods.R
\name{vcov.rd2d}
\alias{vcov.rd2d}
\alias{vcov.rd2d.distance}
\title{Extract rd2d Covariance Matrices}
\usage{
\method{vcov}{rd2d}(object, output = "main", ...)

\method{vcov}{rd2d.distance}(object, output = "main", ...)
}
\arguments{
\item{object}{An object returned by \code{\link{rd2d}} or
\code{\link{rd2d.distance}}.}

\item{output}{Covariance output to extract. Default is \code{"main"}.}

\item{...}{Additional arguments ignored.}
}
\value{
A covariance matrix.
}
\description{
Extract rd2d Covariance Matrices
}
